chore: Handle Bing crawlers separately
Somehow crawlers using this UA run the tracker code,
as they hit a valid URL using fetch, but with the wrong
method: GET instead of POST, and without the payload
in the URL.
This probably means they use a bad version of fetch
which does not handle the method properly. We're adding
a log line to see if they send a body, then we'll assess.